requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-storages==1.13.2
django-tinymce==3.6.1
gunicorn==20.1.0
Pillow==9.5.0
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0